+++ Runtime/Unity.Serialization/Json/Parser/Internal/NodeParser.cs
@@ -1,9 +1,6 @@
 using System;
-using System.Runtime.InteropServices;
-using Unity.Burst;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
-using Unity.Jobs;
 
 namespace Unity.Serialization.Json
 {