+++ Tests/Unity.Serialization.Tests/Json/Adapters/JsonAdapterTests+ReEntrance.cs
@@ -33,13 +33,7 @@ namespace Unity.Serialization.Json.Tests
         {
             public void Serialize(in JsonSerializationContext<Foo> context, Foo value)
             {
-                var json = JsonSerialization.ToJson(value);
-
-                json = json.Replace("\"", "\\\"");
-                json = json.Replace("\n", "");
-                json = json.Replace(" ", "");
-                
-                context.Writer.WriteValue(json);
+                context.Writer.WriteValue(JsonSerialization.ToJson(value).Replace('"', '\"').Replace("\n", "").Replace(" ", ""));
             }
 
             public Foo Deserialize(in JsonDeserializationContext<Foo> context)